Added delete btn for pictures in given route (route-details.html).

+++ project/src/main/java/com/myproject/project/web/PictureController.java
@@ -1,7 +1,6 @@
 package com.myproject.project.web;
 
 import com.myproject.project.model.dto.PictureUploadDto;
-import com.myproject.project.model.dto.PictureViewModel;
 import com.myproject.project.model.entity.RouteEntity;
 import com.myproject.project.model.entity.UserEntity;
 import com.myproject.project.service.PictureService;
@@ -10,12 +9,10 @@ import com.myproject.project.service.UserService;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import javax.transaction.Transactional;
 import java.io.IOException;
-import java.util.List;
 
 @Controller
 public class PictureController {
@@ -43,26 +40,48 @@ public class PictureController {
         //TODO: refactor this part!!!
         if (route.getAuthor().getEmail().equals(author.getEmail())) {
             this.pictureService.addPicture(pictureUploadDto, author, route);
-            return "redirect:/routes/details/" + id;
         } else{
             //don't have permission for this action
-            return "redirect:/routes/details/" + id;
         }
+        return "redirect:/routes/details/" + id;
 
-    }
-
-    @GetMapping("/pictures/all")
-    public String allPictures(Model model) {
-        List<PictureViewModel> pictures = this.pictureService.getAllPictures();
-        model.addAttribute("pictures", pictures);
-        return "all";
     }
 
     @Transactional
-    @DeleteMapping("/pictures/delete")
-    public String delete(@RequestParam("public_id") String publicId){
-        this.pictureService.deleteByPublicId(publicId);
+    @DeleteMapping("/routes/details/{id}/pictures")
+    public String deletePictureForRoute(@PathVariable Long id,
+                                        @AuthenticationPrincipal UserDetails userDetails,
+                                        @RequestParam("public_id") String publicId){
 
-        return "redirect:/pictures/all";
+        UserEntity author = this.userService.findUserByEmail(userDetails.getUsername());
+        RouteEntity route = this.routeService.findRouteEntityById(id);
+
+        boolean isAuthorOfPicture = route.getAuthor().getEmail().equals(author.getEmail())
+                && route.getPictures()
+                .stream()
+                .anyMatch(p -> p.getPublicId().equals(publicId));
+
+        if (isAuthorOfPicture) {
+            this.pictureService.deleteByPublicId(publicId);
+        } else {
+            ////don't have permission for this action
+        }
+
+        return "redirect:/routes/details/" + id;
     }
+
+//    @GetMapping("/pictures/all")
+//    public String allPictures(Model model) {
+//        List<PictureViewModel> pictures = this.pictureService.getAllPictures();
+//        model.addAttribute("pictures", pictures);
+//        return "all";
+//    }
+
+//    @Transactional
+//    @DeleteMapping("/pictures/delete")
+//    public String delete(@RequestParam("public_id") String publicId){
+//        this.pictureService.deleteByPublicId(publicId);
+//
+//        return "redirect:/pictures/all";
+//    }
 }
